fix(backen, main): comment pid saver that was supossed to avoid running the back more than once, does not work properly
